flask/app/requirements.txt
absl-py==2.0.0
aniso8601==9.0.1
anyio==3.7.1
astunparse==1.6.3
attrs==23.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==7.1.2
colorama==0.4.6
cryptography==41.0.5
dnspython==2.1.0
docx==0.2.4
flasgger==0.9.7.1
Flask==1.1.4
Flask-Cors==4.0.0
Flask-JWT-Extended==3.25.0
Flask-Mail==0.9.1
flask-paginate==2023.10.24
Flask-RESTful==0.3.10
flatbuffers==23.5.26
fpdf==1.7.2
gast==0.4.0
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googletrans==3.0.0
greenlet==3.0.1
grpcio==1.59.2
h11==0.9.0
h2==3.2.0
h5py==3.10.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.3.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
keras==2.13.1
libclang==16.0.6
lxml==4.9.3
Markdown==3.5.1
MarkupSafe==2.0.1
mistune==3.0.2
ml-dtypes==0.2.0
mysql==0.0.3
mysql-connector-python==8.2.0
mysqlclient==2.2.0
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
passlib==1.7.4
Pillow==10.1.0
pipdeptree==2.13.1
protobuf==4.21.12
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDocX==0.9.10
PyJWT==1.7.1
pymongo==3.12.1
PyMuPDF==1.23.7
PyMuPDFb==1.23.7
PyPDF2==3.0.1
pytesseract
python-docx==1.1.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rpds-py==0.12.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow
tensorflow-estimator==2.13.0
tensorflow-intel
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.2.0
typing_extensions==4.5.0
urllib3==2.0.7
Werkzeug==1.0.1
wrapt==1.14.1